% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot2pval.R
\name{boot2pval}
\alias{boot2pval}
\title{Bootstrap interval to p-value function}
\usage{
boot2pval(x1, x2, conf.level = 0.95, iter = 3000, mudiff = 0, test = "two")
}
\arguments{
\item{test}{}
}
\value{
a Histogram with the bootstrap interval and the calculated p-value
}
\description{
Function to calculate the bootstrap interval and p-value
}
\examples{
set.seed(10);x=rnorm(20,mean=4,sd=2);y=rnorm(20,mean=4,sd=2)
boot2pval(x1=x,x2=y)
}
