% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{Binoimal experiment function}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{p}{}
}
\value{
a Bar plot with the results from the binomial experiment
}
\description{
Takes in parameters and then runs a binomial experiment
}
\examples{
mybin(iter=1000,n=18, p=0.3)
}
