% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltu.R
\name{mycltu}
\alias{mycltu}
\title{CLT Uniform function}
\usage{
mycltu(n, iter, a = 0, b = 10)
}
\arguments{
\item{b}{}
}
\value{
a Graph with all data needed to find the cetral limit
}
\description{
My Central Limit Function
}
\examples{
mycltu(n=20,iter=100000)
}
