% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myread.R
\name{myread}
\alias{myread}
\title{Read csv function}
\usage{
myread(csv)
}
\arguments{
\item{csv}{}
}
\value{
a Object holding all of the data read from the csv file
}
\description{
Takes a csv file and reads in the data
}
\examples{
File.csv; myread(File.csv)
}
